feat: repeated copy clipboard atom component added to replaced the repeated sub-styled component(s)
